+++ bundles/org.eclipse.passage.lic.api/META-INF/MANIFEST.MF
@@ -9,6 +9,7 @@ Bundle-Copyright: %Bundle-Copyright
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: org.eclipse.passage.lic.api,
  org.eclipse.passage.lic.api.access,
+ org.eclipse.passage.lic.api.acquire,
  org.eclipse.passage.lic.api.conditions,
  org.eclipse.passage.lic.api.conditions.evaluation,
  org.eclipse.passage.lic.api.conditions.mining,
@@ -18,13 +19,6 @@ Export-Package: org.eclipse.passage.lic.api,
  org.eclipse.passage.lic.api.requirements,
  org.eclipse.passage.lic.api.restrictions,
  org.eclipse.passage.lic.api.version,
- org.eclipse.passage.lic.internal.api.acquire;
-  x-friends:="org.eclipse.passage.lic.hc,
-   org.eclipse.passage.lic.floating.model,
-   org.eclipse.passage.lic.base,
-   org.eclipse.passage.lic.hc.tests,
-   org.eclipse.passage.lbc.jetty,
-   org.eclipse.passage.lic.licenses.model",
  org.eclipse.passage.lic.internal.api.diagnostic;
   x-friends:="org.eclipse.passage.lic.base,
    org.eclipse.passage.lic.jface,
