+++ bundles/org.eclipse.passage.lic.api/src/org/eclipse/passage/lic/api/access/GrantLockAttempt.java
@@ -12,8 +12,8 @@
  *******************************************************************************/
 package org.eclipse.passage.lic.api.access;
 
+import org.eclipse.passage.lic.api.acquire.GrantAcquisition;
 import org.eclipse.passage.lic.api.restrictions.ExaminationCertificate;
-import org.eclipse.passage.lic.internal.api.acquire.GrantAcquisition;
 
 /**
  * @since 1.1
