+++ bundles/org.eclipse.passage.lic.api/src/org/eclipse/passage/lic/api/acquire/LicenseAcquisitionServicesRegistry.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 ArSysOp
+ * Copyright (c) 2020, 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -10,23 +10,14 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.internal.base.acquire;
+package org.eclipse.passage.lic.api.acquire;
 
-import java.nio.file.Path;
+import java.util.function.Supplier;
 
 import org.eclipse.passage.lic.api.conditions.ConditionMiningTarget;
+import org.eclipse.passage.lic.api.registry.Registry;
 
-public final class PathLicenseAcquisitionService extends LocalLicenseAcquisitionService {
-
-	private final Path root;
-
-	public PathLicenseAcquisitionService(Path root) {
-		this.root = root;
-	}
-
-	@Override
-	public ConditionMiningTarget id() {
-		return new ConditionMiningTarget.Local().child(root.toAbsolutePath().toString());
-	}
+public interface LicenseAcquisitionServicesRegistry
+		extends Supplier<Registry<ConditionMiningTarget, LicenseAcquisitionService>> {
 
 }
