+++ bundles/org.eclipse.passage.lic.base/META-INF/MANIFEST.MF
@@ -11,16 +11,17 @@ Require-Bundle: org.eclipse.passage.lic.api;bundle-version="1.0.0";visibility:=r
  org.apache.logging.log4j;bundle-version="2.8.2"
 Export-Package: org.eclipse.passage.lic.base,
  org.eclipse.passage.lic.base.access,
+ org.eclipse.passage.lic.base.acquire;x-friends:="org.eclipse.passage.lic.floating.model,org.eclipse.passage.lic.equinox,org.eclipse.passage.lic.licenses.model",
  org.eclipse.passage.lic.base.conditions,
  org.eclipse.passage.lic.base.conditions.evaluation,
  org.eclipse.passage.lic.base.conditions.mining,
+ org.eclipse.passage.lic.base.inspection,
  org.eclipse.passage.lic.base.io,
  org.eclipse.passage.lic.base.registry,
  org.eclipse.passage.lic.base.requirements,
  org.eclipse.passage.lic.base.restrictions,
  org.eclipse.passage.lic.base.version,
  org.eclipse.passage.lic.internal.base;x-friends:="org.eclipse.passage.lic.equinox.tests",
- org.eclipse.passage.lic.internal.base.acquire;x-friends:="org.eclipse.passage.lic.floating.model,org.eclipse.passage.lic.equinox,org.eclipse.passage.lic.licenses.model",
  org.eclipse.passage.lic.internal.base.diagnostic;
   x-friends:="org.eclipse.passage.lbc.base,
    org.eclipse.passage.lic.equinox,
@@ -39,7 +40,6 @@ Export-Package: org.eclipse.passage.lic.base,
    org.eclipse.passage.lbc.base,
    org.eclipse.passage.seal.demo.tests",
  org.eclipse.passage.lic.internal.base.i18n;x-internal:=true,
- org.eclipse.passage.lic.internal.base.inspection;x-friends:="org.eclipse.passage.loc.licenses.core,org.eclipse.passage.lic.oshi",
  org.eclipse.passage.lic.internal.base.inspection.hardware;x-friends:="org.eclipse.passage.loc.licenses.core,org.eclipse.passage.lic.oshi",
  org.eclipse.passage.lic.internal.base.logging;x-friends:="org.eclipse.passage.lic.jetty",
  org.eclipse.passage.lic.internal.base.observatory;x-internal:=true,
