+++ bundles/org.eclipse.passage.lic.base/src/org/eclipse/passage/lic/base/access/BaseGrantLockAttempt.java
@@ -15,8 +15,8 @@ package org.eclipse.passage.lic.base.access;
 import java.util.Objects;
 
 import org.eclipse.passage.lic.api.access.GrantLockAttempt;
+import org.eclipse.passage.lic.api.acquire.GrantAcquisition;
 import org.eclipse.passage.lic.api.restrictions.ExaminationCertificate;
-import org.eclipse.passage.lic.internal.api.acquire.GrantAcquisition;
 
 abstract class BaseGrantLockAttempt implements GrantLockAttempt {
