+++ bundles/org.eclipse.passage.lic.base/src/org/eclipse/passage/lic/base/acquire/BaseGrantAcquisition.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 ArSysOp
+ * Copyright (c) 2020, 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -10,12 +10,12 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.internal.base.acquire;
+package org.eclipse.passage.lic.base.acquire;
 
 import java.util.Date;
 import java.util.Objects;
 
-import org.eclipse.passage.lic.internal.api.acquire.GrantAcquisition;
+import org.eclipse.passage.lic.api.acquire.GrantAcquisition;
 
 public final class BaseGrantAcquisition implements GrantAcquisition {
