+++ bundles/org.eclipse.passage.lic.base/src/org/eclipse/passage/lic/base/inspection/BaseEnvironmentProperty.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 ArSysOp
+ * Copyright (c) 2020, 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -10,7 +10,7 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.internal.base.inspection;
+package org.eclipse.passage.lic.base.inspection;
 
 import java.util.Objects;
