+++ bundles/org.eclipse.passage.lic.equinox/src/org/eclipse/passage/lic/internal/equinox/acquire/ConfigurationLicenseAcquisitionService.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 ArSysOp
+ * Copyright (c) 2020, 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -13,7 +13,7 @@
 package org.eclipse.passage.lic.internal.equinox.acquire;
 
 import org.eclipse.passage.lic.api.conditions.ConditionMiningTarget;
-import org.eclipse.passage.lic.internal.base.acquire.LocalLicenseAcquisitionService;
+import org.eclipse.passage.lic.base.acquire.LocalLicenseAcquisitionService;
 import org.eclipse.passage.lic.internal.equinox.conditions.ConfigurationMiningTarget;
 
 public final class ConfigurationLicenseAcquisitionService extends LocalLicenseAcquisitionService {
