+++ bundles/org.eclipse.passage.lic.licenses.model/plugin.properties
@@ -1,6 +1,6 @@
 #Properties file for org.eclipse.passage.lic.licenses.model
 ###############################################################################
-# Copyright (c) 2018, 2020 ArSysOp and others
+# Copyright (c) 2018, 2021 ArSysOp and others
 #
 # This program and the accompanying materials are made available under the
 # terms of the Eclipse Public License 2.0 which is available at
@@ -14,7 +14,7 @@
 
 pluginName = Passage LIC Licenses Model
 providerName=Eclipse Passage
-Bundle-Copyright = Copyright (c) 2018, 2020 ArSysOp and others.\n\
+Bundle-Copyright = Copyright (c) 2018, 2021 ArSysOp and others.\n\
 \n\
 This program and the accompanying materials are made\n\
 available under the terms of the Eclipse Public License 2.0\n\
