+++ bundles/org.eclipse.passage.lic.licenses.model/src/org/eclipse/passage/lic/internal/licenses/convert/PGrantAcquisition.java
@@ -15,11 +15,11 @@ package org.eclipse.passage.lic.internal.licenses.convert;
 import java.util.Objects;
 import java.util.function.Supplier;
 
-import org.eclipse.passage.lic.internal.base.acquire.BaseGrantAcquisition;
+import org.eclipse.passage.lic.base.acquire.BaseGrantAcquisition;
 import org.eclipse.passage.lic.licenses.model.api.GrantAcqisition;
 
 public final class PGrantAcquisition
-		implements Supplier<org.eclipse.passage.lic.internal.api.acquire.GrantAcquisition> {
+		implements Supplier<org.eclipse.passage.lic.api.acquire.GrantAcquisition> {
 
 	private final GrantAcqisition source;
 
@@ -29,7 +29,7 @@ public final class PGrantAcquisition
 	}
 
 	@Override
-	public org.eclipse.passage.lic.internal.api.acquire.GrantAcquisition get() {
+	public org.eclipse.passage.lic.api.acquire.GrantAcquisition get() {
 		return new BaseGrantAcquisition(//
 				source.getIdentifier(), //
 				source.getGrant(), //
