+++ bundles/org.eclipse.passage.seal.demo/src/org/eclipse/passage/seal/internal/demo/DirectedAccessCycleConfiguration.java
@@ -16,6 +16,8 @@ import java.nio.file.Path;
 import java.util.function.Supplier;
 
 import org.eclipse.passage.lic.api.LicensedProduct;
+import org.eclipse.passage.lic.api.acquire.LicenseAcquisitionService;
+import org.eclipse.passage.lic.api.acquire.LicenseAcquisitionServicesRegistry;
 import org.eclipse.passage.lic.api.conditions.ConditionMiningTarget;
 import org.eclipse.passage.lic.api.conditions.mining.MinedConditions;
 import org.eclipse.passage.lic.api.conditions.mining.MinedConditionsRegistry;
@@ -23,11 +25,9 @@ import org.eclipse.passage.lic.api.io.Hashes;
 import org.eclipse.passage.lic.api.io.HashesRegistry;
 import org.eclipse.passage.lic.api.registry.Registry;
 import org.eclipse.passage.lic.api.registry.StringServiceId;
+import org.eclipse.passage.lic.base.acquire.PathLicenseAcquisitionService;
 import org.eclipse.passage.lic.base.conditions.mining.PathResidentConditions;
 import org.eclipse.passage.lic.base.registry.ReadOnlyRegistry;
-import org.eclipse.passage.lic.internal.api.acquire.LicenseAcquisitionService;
-import org.eclipse.passage.lic.internal.api.acquire.LicenseAcquisitionServicesRegistry;
-import org.eclipse.passage.lic.internal.base.acquire.PathLicenseAcquisitionService;
 import org.osgi.framework.Bundle;
 
 @SuppressWarnings("restriction")
