+++ bundles/org.eclipse.passage.seal.demo/src/org/eclipse/passage/seal/internal/demo/FocusedAccessCycleConfiguration.java
@@ -15,9 +15,9 @@ package org.eclipse.passage.seal.internal.demo;
 import java.util.function.Supplier;
 
 import org.eclipse.passage.lic.api.LicensedProduct;
+import org.eclipse.passage.lic.api.acquire.LicenseAcquisitionServicesRegistry;
 import org.eclipse.passage.lic.api.conditions.mining.MinedConditionsRegistry;
 import org.eclipse.passage.lic.api.io.HashesRegistry;
-import org.eclipse.passage.lic.internal.api.acquire.LicenseAcquisitionServicesRegistry;
 import org.osgi.framework.FrameworkUtil;
 
 @SuppressWarnings("restriction")
