+++ bundles/org.eclipse.passage.seal.demo/src/org/eclipse/passage/seal/internal/demo/PersonalLicensing.java
@@ -15,6 +15,8 @@ package org.eclipse.passage.seal.internal.demo;
 import java.util.Arrays;
 import java.util.function.Supplier;
 
+import org.eclipse.passage.lic.api.acquire.LicenseAcquisitionService;
+import org.eclipse.passage.lic.api.acquire.LicenseAcquisitionServicesRegistry;
 import org.eclipse.passage.lic.api.conditions.ConditionMiningTarget;
 import org.eclipse.passage.lic.api.conditions.mining.MinedConditions;
 import org.eclipse.passage.lic.api.conditions.mining.MinedConditionsRegistry;
@@ -23,11 +25,9 @@ import org.eclipse.passage.lic.api.io.Hashes;
 import org.eclipse.passage.lic.api.io.HashesRegistry;
 import org.eclipse.passage.lic.api.registry.Registry;
 import org.eclipse.passage.lic.api.registry.StringServiceId;
+import org.eclipse.passage.lic.base.acquire.UserHomeLicenseAcquisitionService;
 import org.eclipse.passage.lic.base.conditions.mining.UserHomeResidentConditions;
 import org.eclipse.passage.lic.base.registry.ReadOnlyRegistry;
-import org.eclipse.passage.lic.internal.api.acquire.LicenseAcquisitionService;
-import org.eclipse.passage.lic.internal.api.acquire.LicenseAcquisitionServicesRegistry;
-import org.eclipse.passage.lic.internal.base.acquire.UserHomeLicenseAcquisitionService;
 import org.eclipse.passage.lic.internal.equinox.acquire.ConfigurationLicenseAcquisitionService;
 import org.eclipse.passage.lic.internal.equinox.acquire.InstallationLicenseAcquisitionService;
 import org.eclipse.passage.lic.internal.equinox.conditions.ConfigurationResidentConditions;
