+++ tests/org.eclipse.passage.lic.hc.tests/OSGI-INF/l10n/bundle.properties
@@ -1,6 +1,6 @@
 #Properties file for org.eclipse.passage.lic.hc.tests
 ###############################################################################
-# Copyright (c) 2020 ArSysOp and others
+# Copyright (c) 2020, 2021 ArSysOp and others
 #
 # This program and the accompanying materials are made available under the
 # terms of the Eclipse Public License 2.0 which is available at
@@ -14,7 +14,7 @@
 
 Bundle-Name = Passage LIC HC Tests
 Bundle-Vendor = Eclipse Passage
-Bundle-Copyright = Copyright (c) 2020 ArSysOp and others.\n\
+Bundle-Copyright = Copyright (c) 2020, 2021 ArSysOp and others.\n\
 \n\
 This program and the accompanying materials are made\n\
 available under the terms of the Eclipse Public License 2.0\n\
